feat(trait): All `Predicate`s are now `Display`
This is a step towards #7.  We can now display the expression under
test.

Next steps:
- Ellipsis the strings being reported from this change.
- Ability to enumerate failure cases.

I'm mixed about building this enumeration functionality directly into
`Predcate` or if we should instead provide an expanded trait that is
only implemented when `Predicate` is `Display` and `T` is `Debug`.  For
now, I'm going with the simplest implementation, to assume every
predicate will want this functionality.

BREAKING CHANGES:
- `Display` is now required to implement `Predicate`.
- Generic traits (taking `T`) now require `T` to be `Debug`.